Refactor graphql init function and related classes
