--- src/Operation/Order/AbstractOperation.php
@@ -1,99 +1,0 @@
-<?php
-/**
- * Copyright (c) 2019, Nosto Solutions Ltd
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors
- * may be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2019 Nosto Solutions Ltd
- * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
- *
- */
-
-namespace Nosto\Operation\Order;
-
-use Nosto\Operation\AbstractAuthenticatedOperation;
-
-use Nosto\Nosto;
-use Nosto\NostoException;
-use Nosto\Request\Graphql\GraphqlRequest;
-use Nosto\Request\Api\Token;
-use Nosto\Exception\Builder as ExceptionBuilder;
-
-abstract class AbstractOperation extends AbstractAuthenticatedOperation
-{
-
-    abstract public function getQuery();
-
-    /**
-     * @return GraphqlRequest
-     * @throws NostoException
-     */
-    public function initGraphqlRequest()
-    {
-        $token = $this->account->getApiToken(Token::API_GRAPHQL);
-        if (is_null($token)) {
-            throw new NostoException('No API / Graphql token found for account.');
-        }
-
-        $request = new GraphqlRequest();
-        $request->setResponseTimeout($this->getResponseTimeout());
-        $request->setConnectTimeout($this->getConnectTimeout());
-        $request->setContentType(self::CONTENT_TYPE_APPLICATION_GRAPHQL);
-        $request->setAuthBasic('', $token->getValue());
-        $request->setUrl(Nosto::getGraphqlBaseUrl() . GraphqlRequest::PATH_GRAPH_QL);
-
-        return $request;
-    }
-
-    /**
-     * @throws NostoException
-     */
-    public function execute()
-    {
-        $request = $this->initGraphqlRequest();
-        $response = $request->postRaw(
-            $this->buildPayload()
-        );
-        if ($response->getCode() !== 200) {
-            throw ExceptionBuilder::fromHttpRequestAndResponse($request, $response);
-        }
-
-        return true;
-
-    }
-
-    /**
-     * Removes line breaks from the string
-     *
-     * @return null|string|string[]
-     */
-    public function buildPayload()
-    {
-        return preg_replace('/[\r\n]+/', '', $this->getQuery());
-    }
-}
